health displayed

+++ source/scenes/main_menu.h
@@ -158,7 +158,7 @@ typedef struct{
 // Define the MenuPage struct with an array of MenuPageItem
 typedef struct{
     MenuPageItem items[MAX_MENU_ITEMS];  // Fixed-size array of MenuPageItem
-    size_t itemCount;                    // Number of items currently in the array
+    s32 itemCount;                    // Number of items currently in the array
 	char* pageName;
 } MenuPage;
